refactored Risk analysis samples
